+++ packages/cashc/src/ast/AstTraversal.ts
@@ -1,6 +1,7 @@
 import {
   Node,
   SourceFileNode,
+  ImportNode,
   ContractNode,
   ParameterNode,
   VariableDefinitionNode,
@@ -22,11 +23,13 @@ import {
   ArrayNode,
   TupleIndexOpNode,
   RequireNode,
+  ReturnNode,
   InstantiationNode,
   TupleAssignmentNode,
   NullaryOpNode,
   ConsoleStatementNode,
   ConsoleParameterNode,
+  FunctionCallStatementNode,
   SliceNode,
   DoWhileNode,
   WhileNode,
@@ -36,7 +39,12 @@ import AstVisitor from './AstVisitor.js';
 
 export default class AstTraversal extends AstVisitor<Node> {
   visitSourceFile(node: SourceFileNode): Node {
-    node.contract = this.visit(node.contract) as ContractNode;
+    node.functions = this.visitList(node.functions) as FunctionDefinitionNode[];
+    node.contract = this.visitOptional(node.contract) as ContractNode | undefined;
+    return node;
+  }
+
+  visitImport(node: ImportNode): Node {
     return node;
   }
 
@@ -82,8 +90,18 @@ export default class AstTraversal extends AstVisitor<Node> {
     return node;
   }
 
+  visitReturn(node: ReturnNode): Node {
+    node.expression = this.visit(node.expression);
+    return node;
+  }
+
   visitConsoleStatement(node: ConsoleStatementNode): Node {
     node.parameters = this.visitList(node.parameters) as ConsoleParameterNode[];
+    return node;
+  }
+
+  visitFunctionCallStatement(node: FunctionCallStatementNode): Node {
+    node.functionCall = this.visit(node.functionCall) as FunctionCallNode;
     return node;
   }